+++ examples/symfony-handler.php
@@ -2,12 +2,13 @@
 
 declare(strict_types=1);
 
+use Symfony\Component\HttpFoundation\Request;
 use Tabuna\Map\Mapper;
 
 final class ImportAirportHandler
 {
-    public function __invoke(array $payload): AirportDto
+    public function __invoke(Request $request): AirportDto
     {
-        return Mapper::into($payload, AirportDto::class);
+        return Mapper::into($request, AirportDto::class);
     }
 }